Justify rule duplication
