+++ javascript-checks/src/test/java/org/sonar/javascript/checks/CyclomaticComplexityTypeScriptCheckTest.java
@@ -19,34 +19,19 @@
  */
 package org.sonar.javascript.checks;
 
-import java.util.Collections;
-import java.util.List;
+import org.junit.Test;
 
-import org.sonar.check.Rule;
-import org.sonar.check.RuleProperty;
-import org.sonar.javascript.checks.annotations.JavaScriptRule;
-import org.sonarsource.analyzer.commons.annotations.DeprecatedRuleKey;
+import static org.assertj.core.api.Assertions.assertThat;
 
-@JavaScriptRule
-@DeprecatedRuleKey(ruleKey = "FunctionComplexity")
-@Rule(key = "S1541")
-public class FunctionComplexityCheck extends EslintBasedCheck {
+public class CyclomaticComplexityTypeScriptCheckTest {
 
-  private static final int DEFAULT_THRESHOLD = 10;
-
-  @RuleProperty(
-    key = "maximumFunctionComplexityThreshold",
-    description = "The maximum authorized complexity in function",
-    defaultValue = "" + DEFAULT_THRESHOLD)
-  int threshold = DEFAULT_THRESHOLD;
-
-  @Override
-  public List<Object> configurations() {
-    return Collections.singletonList(threshold);
-  }
-
-  @Override
-  public String eslintKey() {
-    return "cyclomatic-complexity";
+  @Test
+  public void configurations() {
+    CyclomaticComplexityTypeScriptCheck check = new CyclomaticComplexityTypeScriptCheck();
+    // default configuration
+    assertThat(check.configurations()).containsExactly(10);
+    // custom configuration
+    check.threshold = 15;
+    assertThat(check.configurations()).containsExactly(15);
   }
 }
